+++ src/main/java/cart/dao/ProductEntity.java
@@ -1,14 +1,20 @@
-package cart.repository;
+package cart.dao;
 
-public class ProductDto {
+public class ProductEntity {
+    private final Integer id;
     private final String name;
     private final String image;
     private final Long price;
 
-    public ProductDto(final String name, final String image, final Long price) {
+    public ProductEntity(final Integer id, final String name, final String image, final Long price) {
+        this.id = id;
         this.name = name;
         this.image = image;
         this.price = price;
+    }
+
+    public Integer getId() {
+        return id;
     }
 
     public String getName() {
