+++ src/main/java/cart/dao/ProductJdbcDao.java
@@ -1,8 +1,5 @@
 package cart.dao;
 
-import cart.entity.ProductEntity;
-import cart.repository.ProductDto;
-
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
